refactor(api): git remotes 라우트를 전용 모듈로 추출
4개 라우트를 api/git/remotes.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 동일.

의존 체인 승격: get_git_service_for_project 가 get_effective_git_path 에
의존하므로 두 함수를 함께 _shared.py 로 옮겼다. _legacy.py 는
from ._shared import 로 두 이름을 자기 네임스페이스에 다시 바인딩한다 —
tests/backend/api/test_api_git_prune.py 의
patch("api.git._legacy.get_git_service_for_project") 7곳이 계속 유효하다
(7 passed 확인). _legacy.py 상단 models.git import 에서 remotes 전용
심볼 4개(RemoteListResponse·RemoteAddRequest·RemoteUpdateRequest·
RemoteOperationResult)를 제거했다.
