+++ src/backend/api/git/_shared.py
@@ -7,6 +7,8 @@
 
 from fastapi import HTTPException
 
+from models.project import get_project
+
 
 def get_github_service():
     """Get GitHubService instance."""
@@ -18,5 +20,53 @@ def get_github_service():
             status_code=503,
             detail="GitHub service not available. Check GITHUB_TOKEN environment variable.",
         )
+
+    return service
+
+
+def get_effective_git_path(project) -> str:
+    """Get the effective Git path for a project."""
+    return project.git_path or project.path
+
+
+def get_git_service_for_project(project_id: str, worktree_path: str | None = None):
+    """Get GitService for a project, optionally targeting a specific worktree.
+
+    Args:
+        project_id: Project identifier
+        worktree_path: Optional worktree path. Validated against actual worktree list.
+    """
+    from pathlib import Path
+
+    from services.git_service import get_git_service
+
+    project = get_project(project_id)
+    if not project:
+        raise HTTPException(status_code=404, detail=f"Project '{project_id}' not found")
+
+    git_path = get_effective_git_path(project)
+    service = get_git_service(git_path)
+    if not service:
+        raise HTTPException(
+            status_code=400, detail=f"Project '{project_id}' is not a Git repository"
+        )
+
+    if worktree_path:
+        # Security: validate worktree_path against actual worktree list
+        resolved_requested = str(Path(worktree_path).resolve())
+        valid_paths = {str(Path(wt.path).resolve()) for wt in service.list_worktrees()}
+        if resolved_requested not in valid_paths:
+            raise HTTPException(
+                status_code=403,
+                detail="Invalid worktree path: not a registered worktree",
+            )
+        # Return a GitService instance pointing to the worktree
+        wt_service = get_git_service(resolved_requested)
+        if not wt_service:
+            raise HTTPException(
+                status_code=400,
+                detail="Worktree path is not a valid Git working directory",
+            )
+        return wt_service
 
     return service
